+++ apps/example-apps/react/examples/interaction/drag-and-drop/useDnD.tsx
@@ -117,3 +117,23 @@ export const useDnD = () => {
     onDragStart,
   };
 };
+
+export const useDnDPosition = () => {
+  const [position, setPosition] = useState<XYPosition>({ x: 0, y: 0 });
+
+  // By default, the pointer move event sets the position of the dragged element in the context.
+  // This will be used to display the `DragGhost` component.
+  const onDrag = useCallback((event: PointerEvent) => {
+    event.preventDefault();
+    setPosition({ x: event.clientX, y: event.clientY });
+  }, []);
+
+  useEffect(() => {
+    document.addEventListener('pointermove', onDrag);
+    return () => {
+      document.removeEventListener('pointermove', onDrag);
+    };
+  }, [onDrag]);
+
+  return { position };
+};
